+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.5.5
 - Update translations
 - Make ammo index page a bit more compact
+- Add ammo count to ammo type index page
 - Forgot to add the logo as the favicon whoops
 
 # v0.5.4